ml/requirements.txt
fastapi
fastapi[standard]
uvicorn
pillow
torch
torchvision
ultralytics
python-multipart
pytest
scikit-learn
pyyaml